refactor: move the Cell and Script types to src/models
