+++ src/components/TransactionItem/TransactionItemCellList/index.tsx
@@ -2,6 +2,7 @@ import { ReactNode } from 'react'
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import TransactionCellListPanel from './styled'
+import { Cell } from '../../../models/Cell'
 
 const MAX_CELL_SHOW_SIZE = 10
 
@@ -10,9 +11,9 @@ export default ({
   transaction,
   render,
 }: {
-  cells: State.Cell[]
+  cells: Cell[]
   transaction: State.Transaction
-  render: (cell: State.Cell) => ReactNode
+  render: (cell: Cell) => ReactNode
 }) => {
   const { t } = useTranslation()
   return (
