+++ src/models/Script/index.ts
@@ -1,0 +1,5 @@
+export interface Script {
+  codeHash: string
+  args: string
+  hashType: string
+}
