+++ src/pages/Transaction/TransactionCellList/index.tsx
@@ -9,6 +9,7 @@ import { ReactComponent as DeprecatedAddrOn } from '../../../assets/deprecated_a
 import { ReactComponent as DeprecatedAddrOff } from '../../../assets/deprecated_addr_off.svg'
 import { ReactComponent as Warning } from '../../../assets/warning.svg'
 import styles from './styles.module.scss'
+import { Cell } from '../../../models/Cell'
 
 const SCROLL_BOTTOM_OFFSET = 5
 const SCROLL_LOADING_TIME = 400
@@ -20,8 +21,8 @@ export default ({
   showReward,
   addrToggle: { isAddrNew, setIsAddrNew },
 }: {
-  inputs?: State.Cell[]
-  outputs?: State.Cell[]
+  inputs?: Cell[]
+  outputs?: Cell[]
   txHash?: string
   showReward?: boolean
   addrToggle: {
